Refactor starter packages to separate character and plugin configurations
This commit organizes the starter packages to have cleaner separation of concerns:

- **plugin-starter**: Moved plugin configuration to separate plugin.ts file
- **project-starter**: Extracted character configuration to separate character.ts file
- **project-tee-starter**: Updated exports to align with new structure

Benefits:
- Improved code organization and maintainability
- Easier to find and modify character or plugin configurations
- Consistent structure across all starter packages
- Better separation of concerns

All tests pass and builds succeed after refactoring.
