+++ packages/plugin-starter/src/plugin.ts
@@ -1,0 +1,274 @@
+import type { Plugin } from '@elizaos/core';
+import {
+  type Action,
+  type Content,
+  type GenerateTextParams,
+  type HandlerCallback,
+  type IAgentRuntime,
+  type Memory,
+  ModelType,
+  type Provider,
+  type ProviderResult,
+  Service,
+  type State,
+  logger,
+} from '@elizaos/core';
+import { z } from 'zod';
+import { StarterPluginTestSuite } from './tests';
+
+/**
+ * Defines the configuration schema for a plugin, including the validation rules for the plugin name.
+ *
+ * @type {import('zod').ZodObject<{ EXAMPLE_PLUGIN_VARIABLE: import('zod').ZodString }>}
+ */
+const configSchema = z.object({
+  EXAMPLE_PLUGIN_VARIABLE: z
+    .string()
+    .min(1, 'Example plugin variable is not provided')
+    .optional()
+    .transform((val) => {
+      if (!val) {
+        logger.warn('Example plugin variable is not provided (this is expected)');
+      }
+      return val;
+    }),
+});
+
+/**
+ * Example HelloWorld action
+ * This demonstrates the simplest possible action structure
+ */
+/**
+ * Action representing a hello world message.
+ * @typedef {Object} Action
+ * @property {string} name - The name of the action.
+ * @property {string[]} similes - An array of related actions.
+ * @property {string} description - A brief description of the action.
+ * @property {Function} validate - Asynchronous function to validate the action.
+ * @property {Function} handler - Asynchronous function to handle the action and generate a response.
+ * @property {Object[]} examples - An array of example inputs and expected outputs for the action.
+ */
+const helloWorldAction: Action = {
+  name: 'HELLO_WORLD',
+  similes: ['GREET', 'SAY_HELLO'],
+  description: 'Responds with a simple hello world message',
+
+  validate: async (
+    _runtime: IAgentRuntime,
+    _message: Memory,
+    _state: State | undefined
+  ): Promise<boolean> => {
+    // Always valid
+    return true;
+  },
+
+  handler: async (
+    _runtime: IAgentRuntime,
+    message: Memory,
+    _state: State | undefined,
+    _options: any,
+    callback?: HandlerCallback,
+    _responses?: Memory[]
+  ) => {
+    try {
+      logger.info('Handling HELLO_WORLD action');
+
+      // Simple response content
+      const responseContent: Content = {
+        text: 'hello world!',
+        actions: ['HELLO_WORLD'],
+        source: message.content.source,
+      };
+
+      // Call back with the hello world message if callback is provided
+      if (callback) {
+        await callback(responseContent);
+      }
+
+      return responseContent;
+    } catch (error) {
+      logger.error('Error in HELLO_WORLD action:', error);
+      throw error;
+    }
+  },
+
+  examples: [
+    [
+      {
+        name: '{{name1}}',
+        content: {
+          text: 'Can you say hello?',
+        },
+      },
+      {
+        name: '{{name2}}',
+        content: {
+          text: 'hello world!',
+          actions: ['HELLO_WORLD'],
+        },
+      },
+    ],
+  ],
+};
+
+/**
+ * Example Hello World Provider
+ * This demonstrates the simplest possible provider implementation
+ */
+const helloWorldProvider: Provider = {
+  name: 'HELLO_WORLD_PROVIDER',
+  description: 'A simple example provider',
+
+  get: async (
+    _runtime: IAgentRuntime,
+    _message: Memory,
+    _state: State | undefined
+  ): Promise<ProviderResult> => {
+    return {
+      text: 'I am a provider',
+      values: {},
+      data: {},
+    };
+  },
+};
+
+export class StarterService extends Service {
+  static serviceType = 'starter';
+  capabilityDescription =
+    'This is a starter service which is attached to the agent through the starter plugin.';
+  constructor(protected runtime: IAgentRuntime) {
+    super(runtime);
+  }
+
+  static async start(runtime: IAgentRuntime) {
+    logger.info(`*** Starting starter service - MODIFIED: ${new Date().toISOString()} ***`);
+    const service = new StarterService(runtime);
+    return service;
+  }
+
+  static async stop(runtime: IAgentRuntime) {
+    logger.info('*** TESTING DEV MODE - STOP MESSAGE CHANGED! ***');
+    // get the service from the runtime
+    const service = runtime.getService(StarterService.serviceType);
+    if (!service) {
+      throw new Error('Starter service not found');
+    }
+    service.stop();
+  }
+
+  async stop() {
+    logger.info('*** THIRD CHANGE - TESTING FILE WATCHING! ***');
+  }
+}
+
+export const starterPlugin: Plugin = {
+  name: 'plugin-starter',
+  description: 'Plugin starter for elizaOS',
+  config: {
+    EXAMPLE_PLUGIN_VARIABLE: process.env.EXAMPLE_PLUGIN_VARIABLE,
+  },
+  async init(config: Record<string, string>) {
+    logger.info('*** TESTING DEV MODE - PLUGIN MODIFIED AND RELOADED! ***');
+    try {
+      const validatedConfig = await configSchema.parseAsync(config);
+
+      // Set all environment variables at once
+      for (const [key, value] of Object.entries(validatedConfig)) {
+        if (value) process.env[key] = value;
+      }
+    } catch (error) {
+      if (error instanceof z.ZodError) {
+        throw new Error(
+          `Invalid plugin configuration: ${error.errors.map((e) => e.message).join(', ')}`
+        );
+      }
+      throw error;
+    }
+  },
+  models: {
+    [ModelType.TEXT_SMALL]: async (
+      _runtime,
+      { prompt, stopSequences = [] }: GenerateTextParams
+    ) => {
+      return 'Never gonna give you up, never gonna let you down, never gonna run around and desert you...';
+    },
+    [ModelType.TEXT_LARGE]: async (
+      _runtime,
+      {
+        prompt,
+        stopSequences = [],
+        maxTokens = 8192,
+        temperature = 0.7,
+        frequencyPenalty = 0.7,
+        presencePenalty = 0.7,
+      }: GenerateTextParams
+    ) => {
+      return 'Never gonna make you cry, never gonna say goodbye, never gonna tell a lie and hurt you...';
+    },
+  },
+  routes: [
+    {
+      name: 'hello-world-route',
+      path: '/helloworld',
+      type: 'GET',
+      handler: async (_req: any, res: any) => {
+        // send a response
+        res.json({
+          message: 'Hello World!',
+        });
+      },
+    },
+    {
+      name: 'current-time-route',
+      path: '/api/time',
+      type: 'GET',
+      handler: async (_req: any, res: any) => {
+        // Return current time in various formats
+        const now = new Date();
+        res.json({
+          timestamp: now.toISOString(),
+          unix: Math.floor(now.getTime() / 1000),
+          formatted: now.toLocaleString(),
+          timezone: Intl.DateTimeFormat().resolvedOptions().timeZone,
+        });
+      },
+    },
+  ],
+  events: {
+    MESSAGE_RECEIVED: [
+      async (params) => {
+        logger.debug('MESSAGE_RECEIVED event received');
+        // print the keys
+        logger.debug(Object.keys(params));
+      },
+    ],
+    VOICE_MESSAGE_RECEIVED: [
+      async (params) => {
+        logger.debug('VOICE_MESSAGE_RECEIVED event received');
+        // print the keys
+        logger.debug(Object.keys(params));
+      },
+    ],
+    WORLD_CONNECTED: [
+      async (params) => {
+        logger.debug('WORLD_CONNECTED event received');
+        // print the keys
+        logger.debug(Object.keys(params));
+      },
+    ],
+    WORLD_JOINED: [
+      async (params) => {
+        logger.debug('WORLD_JOINED event received');
+        // print the keys
+        logger.debug(Object.keys(params));
+      },
+    ],
+  },
+  services: [StarterService],
+  actions: [helloWorldAction],
+  providers: [helloWorldProvider],
+  tests: [StarterPluginTestSuite],
+  // dependencies: ['@elizaos/plugin-knowledge'], <--- plugin dependecies go here (if requires another plugin)
+};
+
+export default starterPlugin;
